--- promptbench/data/readme.md
@@ -1,12 +1,0 @@
-This folder contains the datasets used for PromptBench. You need to download the datasets you want and put them in this folder.
-
-- Bool_logic: `wget https://wjdcloud.blob.core.windows.net/dataset/promptbench/dataset/bool_logic.json`
-- IWSLT: `wget https://wjdcloud.blob.core.windows.net/dataset/promptbench/dataset/iswst.json`
-- Math: `wget https://wjdcloud.blob.core.windows.net/dataset/promptbench/dataset/math.py`
-- MMLU: `wget https://wjdcloud.blob.core.windows.net/dataset/promptbench/dataset/MMLU.json`
-- MMLU_few_shot: `wget https://wjdcloud.blob.core.windows.net/dataset/promptbench/dataset/MMLU_few_shot.json`
-- SQUAD_V2: `wget https://wjdcloud.blob.core.windows.net/dataset/promptbench/dataset/SQUAD_V2.json`
-- UN_Multi: `wget https://wjdcloud.blob.core.windows.net/dataset/promptbench/dataset/un_multi.json`
-- Valid_parentheses: `wget https://wjdcloud.blob.core.windows.net/dataset/promptbench/dataset/valid_parentheses.json`
-
-Copyright notice: we store all these datasets only for research purpose.
