+++ promptbench/process/postprocess.py
@@ -1,21 +1,6 @@
-def process_pred(dataset_name, pred):
-    if dataset_name in ["cola", "sst2", "mrpc", "qqp", "mnli", "qnli", "rte", "wnli"]:
-        return _process_cls_pred(pred)
-    elif dataset_name == "mmlu":
-        return _process_qa_pred(pred)
-    elif dataset_name == "squad_v2":
-        return _process_squad_v2_pred(pred)
-    elif dataset_name in ['iwslt', 'un_multi']:
-        return _process_trans_pred(pred)
-    elif dataset_name == 'math':
-        return _process_math_pred(pred)
-    elif dataset_name == 'bool_logic':
-        return _process_bool_logic_pred(pred)
-    elif dataset_name == 'valid_parentheses':
-        return _process_valid_parentheses_pred(pred)
-    else:
-        raise NotImplementedError("The dataset is not implemented!")
-
+"""
+TODO: how to process the prediction of different tasks. Using <<<>>>?
+"""
 
 def _process_bool_logic_pred(raw_pred):
     pred = raw_pred.lower()
