+++ promptbench/process/preprocess.py
@@ -1,32 +1,3 @@
-def process_input(prompt, raw_dataset):
-    dataset = raw_dataset.dataset_name
-    if dataset in ["cola", "sst2", "mrpc", "qqp", "mnli", "qnli", "rte", "wnli"]:
-        _process = _process_cls_input
-    elif dataset == "mmlu":
-        _process = _process_qa_input
-    elif dataset == "squad_v2":
-        _process = _process_squad_v2_input
-    elif dataset in ['iwslt', 'un_multi']:
-        _process = _process_trans_input
-    elif dataset == 'math':
-        _process = _process_math_input
-    elif dataset == 'bool_logic':
-        _process = _process_bool_logic_input
-    elif dataset == 'valid_parentheses':
-        _process = _process_valid_parentheses_input
-    else:
-        raise NotImplementedError("The dataset is not implemented!")
-
-    input_texts = []
-    labels = []
-
-    for raw_data in raw_dataset:
-        input_text, label = _process(prompt, raw_data)
-        input_texts.append(input_text)
-        labels.append(label)
-    
-    return input_texts, labels
-
 def _process_valid_parentheses_input(prompt, raw_data):
     question, label = raw_data['question'], raw_data['answer']
     input_text = prompt + '\n' + ("Question: " + question + '\nAnswer: ')
