+++ promptbench/prompts/adv_prompts/chatgpt_zeroshot.md
