+++ promptbench/prompts/adv_prompts/vicuna_fewshot.md
