+++ promptbench/prompts/adv_prompts/vicuna_zeroshot.md
