+++ promptbench/prompts/promptloader.py
@@ -1,7 +1,82 @@
+"""
+TODO: support custom prompt
+"""
+
 import numpy as np
 import re
 import os
 
+"""
+TODO: write a function to add few-shot examples to the prompt
+TODO: evaluation function
+TODO: add code into model.py (there we need to return a score)
+
+TODO: DyVal Llama2?
+TODO: presentation
+"""
+
+def get_few_shot_examples(self):
+    from prompts.three_shot.few_shot_examples import examples
+    few_shot_examples = examples["bool_logic"]
+    return few_shot_examples
+
+
+def get_few_shot_examples(self):
+    from prompts.three_shot.few_shot_examples import examples
+    few_shot_examples = examples["valid_parentheses"]
+    return few_shot_examples
+
+def get_few_shot_examples(self, task):
+    from prompts.three_shot.few_shot_examples import examples
+
+    few_shot_data = examples["math"][task]
+    few_shot_examples = "Here are three examples. \n"
+    for d in few_shot_data:
+        few_shot_examples += "Question: " + d["question"] + "\n"
+        few_shot_examples += "Answer: " + str(d["answer"]) + "\n"
+
+    return few_shot_examples
+
+def get_few_shot_examples(self, task):
+    from prompts.three_shot.few_shot_examples import examples
+
+    few_shot_examples = examples["un_multi"][task]
+    return few_shot_examples
+
+def get_few_shot_examples(self, task):
+    from prompts.three_shot.few_shot_examples import examples
+
+    few_shot_examples = examples["iwslt"][task]
+    return few_shot_examples
+
+def get_few_shot_examples(self, task):
+    from prompts.three_shot.few_shot_examples import examples
+
+    few_shot_examples = examples[task]
+    return few_shot_examples
+
+with open("promptbench/data/MMLU_few_shot.json", "r") as file:
+    self.few_shot_data = json.load(file)
+
+def get_few_shot_examples(self, task):
+    content = "Here are three examples.\n"
+    data = self.few_shot_data[task]
+    for idx in range(min(len(data), 3)):
+        content += ("Input: " + data[idx]["input"] + "\n"
+                    + "A : " + data[idx]["A"] + "\n"
+                    + "B : " + data[idx]["B"] + "\n"
+                    + "C : " + data[idx]["C"] + "\n"
+                    + "D : " + data[idx]["D"] + "\n\n"
+                    + "Answer : " + data[idx]["target"] + "\n"
+                    )
+
+    return content
+
+def get_few_shot_examples(self, task):
+    from prompts.three_shot.few_shot_examples import examples
+
+    few_shot_examples = examples[task]
+    return few_shot_examples
 
 def split_markdown_by_title(markdown_file):
     with open(markdown_file, 'r', encoding='utf-8') as f:
@@ -158,3 +233,5 @@ def retrieve(model_name, dataset_name, attack_name, prompt_type):
 
 if __name__ == "__main__":
     print(retrieve("t5", "cola", "bertattack", "zeroshot_task"))
+
+
